requirements.txt
imapclient
requests
python-dotenv
beautifulsoup4
